서버

+++ Source/Breakout/Game/BOGameInstance.h
@@ -30,8 +30,10 @@ public:
 	ECharacterType GetCharacterType() { return CharacterType; }
 	inline void SetPlayerID(int id) { playerID = id; }
 	inline int GetPlayerID() { return playerID; }
-	inline void SetDeltaTime(double time) { s_deltatime = time; }
-	inline int GetDeltaTime() { return s_deltatime; }
+	inline void SetDeltaTime(double time) {
+		s_deltatime = time;
+	}
+	inline double GetDeltaTime() { return s_deltatime; }
 
 public:
 	class ClientSocket* m_Socket;
